services/image_gen/flux_server/requirements.txt
fastapi
pydantic
diffusers
torch
Pillow
PyYAML
uvicorn
transformers
sentencepiece
peft
gunicorn